ingestao-precifica/requirements.txt
google-cloud-storage==2.16.0
google-cloud-secret-manager==2.20.2
python-dotenv==1.0.1
pandas==2.2.2
requests